% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_func.R
\name{my_function}
\alias{my_function}
\title{My function}
\usage{
my_function(x)
}
\arguments{
\item{x}{numeric, length 1. The number to be squared.}
}
\value{
numeric, length 1. The square of x.
}
\description{
This function takes a number and returns its square.
}
\examples{
my_function(3)
}
